% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{adjacency}
\alias{adjacency}
\title{get weights of all neighbors of a set of nodes}
\usage{
adjacency(x, ...)

adjacency(x, ...)
}
\arguments{
\item{x}{An object representing nearest neighbors.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
An adjacency matrix.
}
\description{
get weights of all neighbors of a set of nodes

Create an adjacency matrix
}
