% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{between_class_neighbors.class_graph}
\alias{between_class_neighbors.class_graph}
\title{Between-Class Neighbors for class_graph Objects}
\usage{
\method{between_class_neighbors}{class_graph}(x, ng, ...)
}
\arguments{
\item{x}{A class_graph object.}

\item{ng}{A neighbor_graph object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A neighbor_graph object representing the between-class neighbors.
}
\description{
Compute the between-class neighbors of a class_graph object.
}
