% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{class_means.class_graph}
\alias{class_means.class_graph}
\title{Class Means for class_graph Objects}
\usage{
\method{class_means}{class_graph}(x, X, ...)
}
\arguments{
\item{x}{A class_graph object.}

\item{X}{The data matrix corresponding to the graph nodes.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A matrix where each row represents the mean values for each class.
}
\description{
Compute the mean of each class for a class_graph object.
}
