% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{convolve_matrix}
\alias{convolve_matrix}
\title{convolve_matrix}
\usage{
convolve_matrix(X, Kern, normalize = FALSE)
}
\arguments{
\item{X}{the data matrix}

\item{Kern}{the Kernel used to smooth the matrix}
}
\description{
convolve_matrix
}
