% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{cross_adjacency}
\alias{cross_adjacency}
\title{Cross Adjacency}
\usage{
cross_adjacency(
  X,
  Y,
  k = 5,
  FUN = heat_kernel,
  type = c("normal", "mutual", "asym"),
  as = c("igraph", "sparse", "index_sim")
)
}
\arguments{
\item{X}{A matrix of size nXk, where n is the number of data points and k is the dimensionality of the feature space.}

\item{Y}{A matrix of size pXk, where p is the number of query points and k is the dimensionality of the feature space.}

\item{k}{An integer indicating the number of nearest neighbors to consider (default: 5).}

\item{FUN}{A kernel function to apply to the Euclidean distances between data points (default: heat_kernel).}

\item{type}{A character string indicating the type of adjacency to compute. One of "normal", "mutual", or "asym" (default: "normal").}

\item{as}{A character string indicating the format of the output. One of "igraph", "sparse", or "index_sim" (default: "igraph").}
}
\value{
If 'as' is "index_sim", a two-column matrix where the first column contains the indices of nearest neighbors and the second column contains the corresponding kernel values.
        If 'as' is "igraph", an igraph object representing the cross adjacency graph.
        If 'as' is "sparse", a sparse adjacency matrix.
}
\description{
This function computes the cross adjacency matrix or graph between two sets of points
based on their k-nearest neighbors and a kernel function applied to their distances.
}
