% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{difference_of_gauss}
\alias{difference_of_gauss}
\title{Compute the Difference of Gaussians for a Coordinate Matrix}
\usage{
difference_of_gauss(
  coord_mat,
  sigma1 = 2,
  sigma2 = sigma1 * 1.6,
  nnk = 3^(ncol(coord_mat))
)
}
\arguments{
\item{coord_mat}{A numeric matrix representing coordinates}

\item{sigma1}{Numeric, the first sigma parameter for the Gaussian smoother (default is 2)}

\item{sigma2}{Numeric, the second sigma parameter for the Gaussian smoother (default is sigma1 * 1.6)}

\item{nnk}{Integer, the number of nearest neighbors for adjacency (default is 3^(ncol(coord_mat)))}
}
\value{
A sparse symmetric matrix representing the computed Difference of Gaussians
}
\description{
Computes the Difference of Gaussians for a given coordinate matrix using specified sigma values and the number of nearest neighbors.
}
\details{
The function first computes two spatial smoother matrices using the specified sigma values,
and then subtracts the second matrix from the first.
}
\examples{
# Create an example coordinate matrix
coord_mat <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 3, ncol = 2)

# Compute the Difference of Gaussians with sigma1 = 2 and sigma2 = 3.2
result <- difference_of_gauss(coord_mat, sigma1 = 2, sigma2 = 3.2)

}
