% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discriminating_similarity.R
\name{discriminating_simililarity2}
\alias{discriminating_simililarity2}
\title{Compute Discriminating Similarity}
\usage{
discriminating_simililarity2(X, labels, k = nrow(X)/2, sigma = 1)
}
\arguments{
\item{X}{Data matrix}

\item{labels}{Factor of class labels}

\item{k}{Number of nearest neighbors}

\item{sigma}{Bandwidth parameter for heat kernel}
}
\value{
A similarity matrix emphasizing discriminating features
}
\description{
Computes a similarity matrix that emphasizes discriminating features between classes.
}
