% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{dist_to_sim.Matrix}
\alias{dist_to_sim.Matrix}
\title{Convert Distance Matrix to Similarity Matrix}
\usage{
\method{dist_to_sim}{Matrix}(
  x,
  method = c("heat", "binary", "normalized", "cosine", "correlation"),
  sigma = 1,
  len = 1
)
}
\arguments{
\item{x}{A sparse Matrix object containing distances}

\item{method}{Character string specifying the transformation method}

\item{sigma}{Scaling parameter for heat kernel transformation}

\item{len}{Feature dimensionality for normalization}
}
\value{
Modified Matrix with distances converted to similarities
}
\description{
Converts a sparse distance matrix to a similarity matrix using various transformation methods.
}
