% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{dist_to_sim.nn_search}
\alias{dist_to_sim.nn_search}
\title{Convert Distances to Similarities}
\usage{
\method{dist_to_sim}{nn_search}(
  x,
  method = c("heat", "binary", "normalized", "cosine", "correlation"),
  sigma = 1
)
}
\arguments{
\item{x}{An nn_search object containing distances}

\item{method}{Character string specifying the transformation method}

\item{sigma}{Scaling parameter for heat kernel transformation}
}
\value{
Modified nn_search object with distances converted to similarities
}
\description{
Converts distance measures to similarity scores using various transformation methods.
}
\details{
Available transformation methods:
* heat: exp(-d^2/sigma)
* binary: 1 if connected, 0 otherwise
* normalized: 1 - d/max(d)
* cosine: cos(d)
* correlation: correlation coefficient
}
