% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{dist_to_sim}
\alias{dist_to_sim}
\title{Convert Distance to Similarity}
\usage{
dist_to_sim(x, ...)
}
\arguments{
\item{x}{An object representing distances (Matrix, nn_search, etc.).}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A similarity matrix or object with distances converted to similarities.
}
\description{
Convert distance values to similarity values using various transformation methods.
}
