% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_zoom.R
\name{.compute_degree_matrix}
\alias{.compute_degree_matrix}
\title{GraphZoom: Faithful Implementation with RcppHNSW-based kNN}
\usage{
.compute_degree_matrix(A)
}
\description{
This code implements the GraphZoom framework as per the given paper:
"GraphZoom: A Multi-level Spectral Approach for Accurate and Scalable Graph Embedding"
}
\details{
Added Feature:
- You can now provide a precomputed feature graph (A_feat). If provided, it will be used directly.
  Otherwise, if X is provided, it will construct the feature graph from X. If neither is provided,
  it will rely solely on the topology graph.

The four phases are:
1) Graph Fusion (Sec. 3.1)
2) Spectral Graph Coarsening (Sec. 3.2)
3) Graph Embedding on the Coarsest Graph (Sec. 3.3)
4) Embedding Refinement (Sec. 3.4)
}
\references{
GraphZoom paper; RcppHNSW documentation
}
