% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{.create_custom_matrix}
\alias{.create_custom_matrix}
\title{Create a Custom Label Matrix (Internal)}
\usage{
.create_custom_matrix(a, b, type = c("s", "d"), simfun = NULL)
}
\description{
Create a Custom Label Matrix (Internal)
}
\keyword{internal}
