% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{.create_dense_matrix}
\alias{.create_dense_matrix}
\title{Create a Dense Label Matrix (Internal)}
\usage{
.create_dense_matrix(a, b, type = c("s", "d"), simfun = NULL)
}
\description{
Create a Dense Label Matrix (Internal)
}
\keyword{internal}
