% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{.create_sparse_matrix}
\alias{.create_sparse_matrix}
\title{Create a Sparse Label Matrix (Internal)}
\usage{
.create_sparse_matrix(a, b = NULL, type = c("s", "d"))
}
\description{
Create a Sparse Label Matrix (Internal)
}
\keyword{internal}
