% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{edges.neighbor_graph}
\alias{edges.neighbor_graph}
\title{Edges for neighbor_graph Objects}
\usage{
\method{edges}{neighbor_graph}(x)
}
\arguments{
\item{x}{A neighbor_graph object.}
}
\value{
A matrix containing the edges of the neighbor_graph object.
}
\description{
Retrieve the edges of a neighbor_graph object.
}
