% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{edges}
\alias{edges}
\title{Edges for Graph-Like Objects}
\usage{
edges(x, ...)
}
\arguments{
\item{x}{A graph-like object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A matrix containing the edges of the graph-like object.
}
\description{
Retrieve the edges of a graph-like object.
}
