% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{expand_label_similarity}
\alias{expand_label_similarity}
\title{Expand Similarity Between Labels Based on a Precomputed Similarity Matrix}
\usage{
expand_label_similarity(labels, sim_mat, threshold = 0, above = TRUE)
}
\arguments{
\item{labels}{A vector of labels for which the similarities will be expanded.}

\item{sim_mat}{A precomputed similarity matrix containing similarities between the unique labels.}

\item{threshold}{A threshold value used to filter the expanded similarity values (default: 0).}

\item{above}{A boolean flag indicating whether to include the values above the threshold (default: TRUE) or below the threshold (FALSE).}
}
\value{
A sparse symmetric similarity matrix with the expanded similarity values.
}
\description{
Expands the similarity between labels based on a precomputed similarity matrix, `sim_mat`, with either above-threshold or below-threshold values depending on the value of the `above` parameter.
}
