% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_constraints.R
\name{feature_weighted_spatial_constraints}
\alias{feature_weighted_spatial_constraints}
\title{feature weighted spatial constraints}
\usage{
feature_weighted_spatial_constraints(
  coords,
  feature_mats,
  sigma_within = 5,
  sigma_between = 3,
  wsigma_within = 0.73,
  wsigma_between = 0.73,
  alpha_within = 0.5,
  alpha_between = 0.5,
  shrinkage_factor = 0.1,
  nnk_within = 27,
  nnk_between = 27,
  maxk_within = nnk_within,
  maxk_between = nnk_between,
  weight_mode_within = "heat",
  weight_mode_between = "binary",
  variable_weights = rep(1, ncol(coords) * length(feature_mats)),
  verbose = FALSE
)
}
\description{
feature weighted spatial constraints
}
\examples{

coords <- as.matrix(expand.grid(1:10, 1:10))
fmats <- replicate(20, matrix(rnorm(100*10), 10, 100), simplify=FALSE)
conmat <- feature_weighted_spatial_constraints(coords, fmats)

conmat <- feature_weighted_spatial_constraints(coords, fmats, maxk_between=4, maxk_within=2,sigma_between=5, nnk_between=60)

#future::plan("multiprocess")
#conmat <- feature_weighted_spatial_constraints(coords, fmats)
}
