% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{find_nn.nnsearcher}
\alias{find_nn.nnsearcher}
\title{Find Nearest Neighbors}
\usage{
\method{find_nn}{nnsearcher}(x, query = NULL, k = 5)
}
\arguments{
\item{x}{A nnsearcher object}

\item{query}{Optional query matrix for searching external points}

\item{k}{Number of nearest neighbors to find}
}
\value{
An nn_search object containing:
* idx: Matrix of indices to nearest neighbors
* dist: Matrix of distances to nearest neighbors
* labels: Matrix of labels of nearest neighbors
}
\description{
Performs nearest neighbor search using a pre-built HNSW index.
}
\examples{
# Search in iris dataset
data(iris)
X <- as.matrix(iris[, 1:4])
searcher <- nnsearcher(X)

# Find 5 nearest neighbors for all points
nn <- find_nn(searcher, k = 5)

# Search for specific query points
query <- matrix(rnorm(8), 2, 4)
nn_query <- find_nn(searcher, query = query, k = 3)

}
