% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{find_nn_among.nn_searcher}
\alias{find_nn_among.nn_searcher}
\title{Find Nearest Neighbors Among Subset}
\usage{
\method{find_nn_among}{nn_searcher}(x, k = 5, idx)
}
\arguments{
\item{x}{A nn_searcher object}

\item{k}{Number of nearest neighbors to find}

\item{idx}{Indices of points to search among}
}
\value{
An nn_search object for the subset
}
\description{
Performs nearest neighbor search within a specified subset of points.
}
