% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{find_nn_between.nn_searcher}
\alias{find_nn_between.nn_searcher}
\title{Find Nearest Neighbors Between Groups}
\usage{
\method{find_nn_between}{nn_searcher}(x, k = 5, idx1, idx2, restricted = FALSE)
}
\arguments{
\item{x}{A nn_searcher object}

\item{k}{Number of nearest neighbors to find}

\item{idx1}{Indices of first group}

\item{idx2}{Indices of second group}

\item{restricted}{Logical; if TRUE, restricts search to specified indices}
}
\value{
An nn_search object containing between-group neighbors
}
\description{
Performs nearest neighbor search between two specified groups of points.
}
