% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{heterogeneous_neighbors}
\alias{heterogeneous_neighbors}
\title{Heterogeneous Neighbors for class_graph Objects}
\usage{
heterogeneous_neighbors(x, X, k, weight_mode = "binary", sigma = 1, ...)
}
\arguments{
\item{x}{A class_graph object.}

\item{X}{The data matrix corresponding to the graph nodes.}

\item{k}{The number of nearest neighbors to find.}

\item{weight_mode}{Method for weighting edges (e.g., "heat", "binary", "euclidean").}

\item{sigma}{Scaling factor for heat kernel if `weight_mode="heat"`.}

\item{...}{Additional arguments passed to weight function.}
}
\value{
A neighbor_graph object representing the between-class neighbors.
}
\description{
Compute the neighbors between different classes for a class_graph object.
}
