% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{indices_to_sparse}
\alias{indices_to_sparse}
\title{internal function}
\usage{
indices_to_sparse(nn.index, hval, return_triplet = FALSE,
  idim = nrow(nn.index), jdim = nrow(nn.index))
}
\description{
internal function
}
