% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{label_matrix}
\alias{label_matrix}
\title{Compute a Similarity or Distance Matrix for Categorical Labels}
\usage{
label_matrix(
  a,
  b,
  type = c("s", "d"),
  return_matrix = TRUE,
  dim1 = length(a),
  dim2 = length(b)
)
}
\arguments{
\item{a}{The first categorical label vector.}

\item{b}{The second categorical label vector.}

\item{type}{The type of matrix to construct, either a similarity ('s') or distance ('d') matrix.}

\item{return_matrix}{A logical flag indicating whether to return the result as a sparse matrix (default: TRUE) or a triplet.}

\item{dim1}{The length of the first dimension.}

\item{dim2}{The length of the second dimension.}
}
\value{
A similarity or distance matrix between the categorical label vectors `a` and `b`, either as a sparse matrix or a triplet.
}
\description{
Calculates a similarity or distance matrix between two categorical label vectors `a` and `b`.
}
