% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{label_sim_matrix}
\alias{label_sim_matrix}
\title{label_sim_matrix}
\usage{
label_sim_matrix(a, b, type = c("s", "d"), return_matrix = TRUE)
}
\arguments{
\item{a}{the first categorical label vector}

\item{b}{the second categorical label vector}

\item{type}{construct a a similarity ('s') or distance ('d') matrix.}

\item{return_matrix}{return as a sparse matrix or a triplet.}
}
\description{
label_sim_matrix
}

