% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{laplacian}
\alias{laplacian}
\title{Compute Graph Laplacian of a Weight Matrix}
\usage{
laplacian(x, ...)
}
\arguments{
\item{x}{The weight matrix representing the graph structure.}

\item{...}{Additional arguments to be passed to specific implementations of the laplacian method.}
}
\value{
A matrix representing the graph Laplacian of the input weight matrix.
}
\description{
This function computes the graph Laplacian of a given weight matrix. The graph Laplacian is defined as 
the difference between the degree matrix and the adjacency matrix.
}
\examples{
# Create a simple weight matrix
w <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), nrow = 3)
L <- laplacian(w)

}
