% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{make_doubly_stochastic}
\alias{make_doubly_stochastic}
\title{make_doubly_stochastic}
\usage{
make_doubly_stochastic(A, iter = 15)
}
\arguments{
\item{A}{the smoothing matrix}

\item{iter}{number of iterations}
}
\description{
make_doubly_stochastic
}
