% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{nclasses.class_graph}
\alias{nclasses.class_graph}
\title{Get Number of Classes in a class_graph Object}
\usage{
\method{nclasses}{class_graph}(x)
}
\arguments{
\item{x}{A class_graph object.}
}
\value{
An integer indicating the number of unique classes in the graph.
}
\description{
Get Number of Classes in a class_graph Object
}
\examples{
data(iris)
cg <- class_graph(iris$Species)
n <- nclasses(cg)  # Should return 3 for iris dataset

}
