% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R, R/neighbor_graph.R
\name{neighbor_graph}
\alias{neighbor_graph}
\alias{neighbor_graph.igraph}
\alias{neighbor_graph.Matrix}
\title{Neighbor Graph}
\usage{
neighbor_graph(x, ...)

\method{neighbor_graph}{igraph}(x, params = list(), type = NULL, classes = NULL, ...)

\method{neighbor_graph}{Matrix}(x, params = list(), type = NULL, classes = NULL, ...)
}
\arguments{
\item{x}{An igraph or Matrix object.}

\item{...}{Additional arguments.}

\item{params}{A list of parameters (default: empty list).}

\item{type}{A character string specifying the type of graph (currently unused).}

\item{classes}{A character vector specifying additional classes for the object.}
}
\value{
A neighbor_graph object, the structure of which depends on the input object's class.

A neighbor_graph object wrapping the input graph structure.
}
\description{
A generic function to create a neighbor_graph object.

Create a neighbor_graph object from an igraph object or Matrix object.
}
\examples{
# Create from igraph
library(igraph)
g <- make_ring(5)
ng1 <- neighbor_graph(g)

# Create from Matrix
adj_matrix <- Matrix::Matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 0), 
                            nrow = 3, byrow = TRUE, sparse = TRUE)
ng2 <- neighbor_graph(adj_matrix)

}
