% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{neighbors}
\alias{neighbors}
\title{Get Neighbors of Graph Vertices}
\usage{
neighbors(x, i, ...)
}
\arguments{
\item{x}{The graph or graph-like object in which to find the neighbors.}

\item{i}{The vertex or vertices for which to find the neighbors. Can be a single vertex 
index or a vector of vertex indices.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A list where each element contains the indices of neighbors for the corresponding 
        input vertex. The length of the list equals the length of input `i`.
}
\description{
This function retrieves the indices of neighbors for one or more vertices in a given graph 
or graph-like object. It is a generic method that can work with various graph representations.
}
\examples{
# Create an example graph
library(igraph)
g <- neighbor_graph(make_ring(5))

# Get neighbors of vertex 1
n <- neighbors(g, 1)

# Get neighbors of multiple vertices
n_multi <- neighbors(g, c(1, 2))

}
