% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{new_repulsion_graph}
\alias{new_repulsion_graph}
\title{Constructor for Repulsion Graph Objects (Internal)}
\usage{
new_repulsion_graph(adjacency_matrix, params = list(), ...)
}
\arguments{
\item{adjacency_matrix}{The final sparse adjacency matrix of the repulsion graph.}

\item{params}{A list of parameters used in its construction.}

\item{...}{Additional attributes to store (inherited from neighbor_graph).}
}
\value{
An object of class `c("repulsion_graph", "neighbor_graph")`.
}
\description{
Constructor for Repulsion Graph Objects (Internal)
}
\keyword{internal}
