% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{node_density}
\alias{node_density}
\title{Node Density}
\usage{
node_density(x, ...)
}
\arguments{
\item{x}{A graph-like object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector of node densities.
}
\description{
Compute the local density around each node in a graph.
}
