% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{non_neighbors.neighbor_graph}
\alias{non_neighbors.neighbor_graph}
\title{Non-neighbors for neighbor_graph Objects}
\usage{
\method{non_neighbors}{neighbor_graph}(x, i)
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{i}{The index of the node for which non-neighboring nodes will be returned.}
}
\value{
A list of node indices that are not neighbors of the given node.
}
\description{
Retrieve the non-neighboring nodes of a given node in a neighbor_graph object.
}
