% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{non_neighbors}
\alias{non_neighbors}
\title{get indices of the non-neighbors of a set of nodes}
\usage{
non_neighbors(x, ...)
}
\arguments{
\item{x}{the neighbor graph}

\item{...}{extra args}
}
\description{
get indices of the non-neighbors of a set of nodes
}
