% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{normalize_adjacency}
\alias{normalize_adjacency}
\title{Normalize Adjacency Matrix}
\usage{
normalize_adjacency(sm, symmetric = TRUE)
}
\arguments{
\item{sm}{A sparse adjacency matrix representing the graph.}

\item{symmetric}{A logical value indicating whether to symmetrize the matrix after normalization (default: TRUE).}
}
\value{
A normalized and, if requested, symmetrized adjacency matrix.
}
\description{
Normalizes an adjacency matrix by dividing each element by the product of the square root of the corresponding row and column sums.
}
\details{
Optionally, it can also symmetrize the normalized matrix by averaging it with its transpose.
}
