% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{normalized_heat_kernel}
\alias{normalized_heat_kernel}
\title{normalized_heat_kernel}
\usage{
normalized_heat_kernel(x, sigma = 1, len)
}
\arguments{
\item{x}{the distances}

\item{sigma}{the bandwidth}

\item{len}{the normalization factor (e.g. the length of the feature vectors)}
}
\description{
normalized_heat_kernel
}

