% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{nvertices.neighbor_graph}
\alias{nvertices.neighbor_graph}
\title{Get number of vertices in neighbor_graph object}
\usage{
\method{nvertices}{neighbor_graph}(x, ...)
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
An integer representing the number of vertices in the graph.
}
\description{
Get the number of vertices in a neighbor_graph object.
}
\examples{
adj_matrix <- Matrix::Matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 0), 
                            nrow = 3, byrow = TRUE, sparse = TRUE)
ng <- neighbor_graph(adj_matrix)
nvertices(ng)  # Should return 3

}
