% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{nvertices}
\alias{nvertices}
\title{Number of Vertices in Graph-like Objects}
\usage{
nvertices(x, ...)
}
\arguments{
\item{x}{an object with a neighborhood}

\item{...}{Additional arguments (currently ignored).}
}
\value{
The number of vertices in the neighbor_graph object.
}
\description{
Retrieve the number of vertices in a neighbor_graph object.
}
\examples{
# Create a small example graph
adj_matrix <- matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 0), nrow = 3, byrow = TRUE)
ng <- neighbor_graph(adj_matrix)
nvertices(ng) # Should return 3

}
