% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{print.class_graph}
\alias{print.class_graph}
\title{Print method for class_graph objects}
\usage{
\method{print}{class_graph}(x, ...)
}
\arguments{
\item{x}{A class_graph object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for class_graph objects
}
