% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{print.neighbor_graph}
\alias{print.neighbor_graph}
\title{Print method for neighbor_graph objects}
\usage{
\method{print}{neighbor_graph}(x, ...)
}
\arguments{
\item{x}{A neighbor_graph object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for neighbor_graph objects
}
