% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{print.repulsion_graph}
\alias{print.repulsion_graph}
\title{Print method for repulsion_graph objects}
\usage{
\method{print}{repulsion_graph}(x, ...)
}
\arguments{
\item{x}{A repulsion_graph object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for repulsion_graph objects
}
