% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{search_result.nnsearcher}
\alias{search_result.nnsearcher}
\title{Format Search Results}
\usage{
\method{search_result}{nnsearcher}(x, result)
}
\arguments{
\item{x}{A nnsearcher object}

\item{result}{Raw search results from HNSW}
}
\value{
A formatted search result object of class "nn_search"
}
\description{
Formats the results from a nearest neighbor search, adding metadata about the
feature dimensionality and distance metric used.
}
