% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{search_result}
\alias{search_result}
\title{Search result for nearest neighbor search}
\usage{
search_result(x, result)
}
\arguments{
\item{x}{An object of class "nnsearcher".}

\item{result}{The result from the nearest neighbor search.}
}
\value{
An object with the class "nn_search".
}
\description{
Search result for nearest neighbor search
}
\examples{
# Create an object of class "nnsearcher"
nn <- nnsearcher()

# Perform nearest neighbor search
result <- find_nn(nn)

# Create search result object
search_result <- search_result(nn, result)

}
