% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{sim_from_adj}
\alias{sim_from_adj}
\title{sim_from_adj}
\usage{
sim_from_adj(A, k = 5, type = c("normal", "mutual"), ncores = 1)
}
\arguments{
\item{A}{adjacency matrix}

\item{k}{number of neighbors}

\item{type}{'normal' or 'mutual' nearest neighbors}

\item{ncores}{number of cores to use}
}
\description{
extract the k-nearest neighbors from an existing adjacency matrix
}
