% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{sim_knn_from_adj}
\alias{sim_knn_from_adj}
\title{sim_knn_from_adj}
\usage{
sim_knn_from_adj(A, k = 5, type = c("normal", "mutual"), ncores = 1)
}
\arguments{
\item{A}{adjacency matrix}

\item{k}{numbe rof neighbors}

\item{type}{normal or mutual nearest neighbors}

\item{ncores}{number of cores to use}
}
\description{
sim_knn_from_adj
}

