% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{similarity_matrix}
\usage{
similarity_matrix(X, neighbor_mode = c("knn", "supervised"),
  weight_mode = c("heat", "normalized", "binary"), k = 5, sigma = 1,
  labels = NULL)
}
\arguments{
\item{X}{the data matrix, where each row is an instance and each column is a variable.}

\item{neighbor_mode}{the method for assigning weights to neighbors}

\item{weight_mode}{binary (1 if neighbor, 0 otherwise), heat kernel, normalized heat kernel}

\item{k}{number of neighbors}

\item{sigma}{parameter for heat kernel \code{exp(-dist/(2*sigma^2))}}

\item{labels}{the class of the categories when \code{weight_mode} is \code{supervised}, supplied as a \code{factor} with \code{nrow(labels) == nrow(X)}}
}
\description{
similarity_matrix
}
