% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{spatial_lap_of_gauss}
\alias{spatial_lap_of_gauss}
\title{spatial_lap_of_gauss}
\usage{
spatial_lap_of_gauss(coord_mat, sigma = 2)
}
\arguments{
\item{coord_mat}{the matrix of coordinates}

\item{sigma}{the standard deviation of the spatial smoother}
}
\description{
spatial_lap_of_gauss
}

