% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{weighted_knnx}
\alias{weighted_knnx}
\title{weighted_knn}
\usage{
weighted_knnx(X, query, k = 5, FUN = heat_kernel, type = c("normal",
  "mutual", "asym"), return_triplet = FALSE)
}
\arguments{
\item{X}{the data matrix, where rows are instances and columns are features}

\item{FUN}{the function used to convert euclidean distances to similarities}

\item{eps}{the e-radius used to threshold neighbor inclusion}
}
\description{
weighted_knn
}
