% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{weighted_spatial_adjacency}
\alias{weighted_spatial_adjacency}
\title{weighted_spatial_adjacency}
\usage{
weighted_spatial_adjacency(coord_mat, feature_mat, wsigma = 0.73,
  alpha = 0.5, nnk = 27, weight_mode = c("binary", "heat"), sigma = 1,
  dthresh = sigma * 2.5, include_diagonal = TRUE, normalized = TRUE,
  stochastic = FALSE)
}
\arguments{
\item{coord_mat}{the coordinate matrix}

\item{feature_mat}{the feature matrix, where: (\code{nrow(feature_mat) == nrow(coord_mat)})}

\item{wsigma}{the sigma for the feature heat kernel}

\item{alpha}{the mixing weight for the spatial distance (1=all spatial, 0=all feature)}

\item{dthresh}{the threshold for the spatial distance}

\item{normalized}{whether to normalize the rows to sum to 1}
}
\description{
weighted_spatial_adjacency
}

