% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{within_class_neighbors}
\alias{within_class_neighbors}
\title{Within-Class Neighbors}
\usage{
within_class_neighbors(x, ng, ...)
}
\arguments{
\item{x}{An object.}

\item{ng}{A neighbor graph object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
An object representing the within-class neighbors of the input graph, the structure of which depends on the input object's class.
}
\description{
A generic function to compute the within-class neighbors of a graph.
}
\examples{
# Create an example graph
g <- neighbor_graph(make_ring(5))

# Compute within-class neighbors
within_n <- within_class_neighbors(g, ng)

}
