% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{adjacency.neighbor_graph}
\alias{adjacency.neighbor_graph}
\title{Adjacency Matrix for neighbor_graph Objects}
\usage{
\method{adjacency}{neighbor_graph}(x, attr = "weight")
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{attr}{A character string specifying the attribute to use for the adjacency matrix.}
}
\value{
A Matrix object representing the adjacency matrix.
}
\description{
Extract the adjacency matrix from a neighbor_graph object.
}
