% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher.R
\name{adjacency.nnsearch}
\alias{adjacency.nnsearch}
\title{Create Adjacency Matrix from nnsearch Object}
\usage{
\method{adjacency}{nnsearch}(x, idim = nrow(x$idx), jdim = max(x$idx), return_triplet = FALSE)
}
\arguments{
\item{x}{An object of class "nnsearch".}

\item{idim}{The number of rows in the resulting matrix.}

\item{jdim}{The number of columns in the resulting matrix.}

\item{return_triplet}{Logical; whether to return triplet format.}
}
\value{
A sparse Matrix representing the adjacency matrix.
}
\description{
Convert a nearest neighbor search result to a sparse adjacency matrix.
}
