% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{adjacency}
\alias{adjacency}
\title{get weights of all neighbors of a set of nodes}
\usage{
adjacency(x, ...)
}
\arguments{
\item{x}{the neighbor graph}
}
\description{
get weights of all neighbors of a set of nodes
}
