% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{between_class_neighbors.class_graph}
\alias{between_class_neighbors.class_graph}
\title{Between-Class Neighbors for class_graph Objects}
\usage{
\method{between_class_neighbors}{class_graph}(x, ng, ...)
}
\arguments{
\item{x}{A class_graph object containing class membership information.}

\item{ng}{A neighbor_graph object containing the base network structure.}

\item{...}{Additional arguments passed to methods (currently unused).}
}
\value{
A neighbor_graph object containing only between-class connections.
}
\description{
Computes a neighbor graph that only includes connections between vertices of different classes.
This is useful for analyzing relationships between different classes in the network.
}
\examples{
# Create example data
data(iris)
X <- as.matrix(iris[, 1:4])
cg <- class_graph(iris$Species)
ng <- neighbor_graph(X)
between_ng <- between_class_neighbors(cg, ng)

}
