% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{bilateral_smoother}
\alias{bilateral_smoother}
\title{bilateral spatial smoother}
\usage{
bilateral_smoother(
  coord_mat,
  feature_mat,
  nnk = 27,
  s_sigma = 2.5,
  f_sigma = 0.7,
  stochastic = FALSE
)
}
\arguments{
\item{coord_mat}{the matrix of coordinates}

\item{nnk}{the number of neighbors in the kernel}

\item{s_sigma}{spatial bandwidth in standard deviations}

\item{f_sigma}{normalized feature bandwidth in standard deviations}

\item{stochastic}{make doubly stochastic}
}
\description{
bilateral spatial smoother
}
\examples{

coord_mat = as.matrix(expand.grid(1:10, 1:10))
feature_mat = matrix(rnorm(100*10), 100, 10)

S <- bilateral_smoother(coord_mat, feature_mat, nnk=8)

}
