% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{binary_label_matrix}
\alias{binary_label_matrix}
\title{create a binary adjaceny matrix}
\usage{
binary_label_matrix(a, b = NULL, type = c("s", "d"))
}
\arguments{
\item{a}{the first categorical label vector}

\item{b}{the second categorical label vector}

\item{type}{construct a a similarity ('s') or distance ('d') matrix.}
}
\description{
Given a set of labels `a` and `b`, construct a binary adjacency matrix containing edges when `a==b`.
}
\examples{
data(iris)

a <- iris[,5]
bl <- binary_label_matrix(a, type="d")
}
