% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{class_graph}
\alias{class_graph}
\title{construct a class_graph}
\usage{
class_graph(labels, sparse = TRUE)
}
\arguments{
\item{labels}{the class label vector}
}
\description{
A graph in which members of the same class have edges
}
\examples{
data(iris)
labels <- iris[,5]
cg <- class_graph(labels)

}
