% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{class_graph}
\alias{class_graph}
\title{Construct a Class Graph}
\usage{
class_graph(labels, sparse = TRUE)
}
\arguments{
\item{labels}{A vector of class labels.}

\item{sparse}{A logical value, indicating whether to use sparse matrices in the computation. Default is TRUE.}
}
\value{
A class_graph object, which is a list containing the following components:
\itemize{
  \item{adjacency}{A matrix representing the adjacency of the graph.}
  \item{params}{A list of parameters used in the construction of the graph.}
  \item{labels}{A vector of class labels.}
  \item{class_indices}{A list of vectors, each containing the indices of elements belonging to a specific class.}
  \item{class_freq}{A table of frequencies for each class.}
  \item{levels}{A vector of unique class labels.}
  \item{classes}{A character string indicating the type of graph ("class_graph").}
}
}
\description{
A graph in which members of the same class have edges.
}
\examples{
data(iris)
labels <- iris[,5]
cg <- class_graph(labels)

}
