% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{class_means.class_graph}
\alias{class_means.class_graph}
\title{Compute Class Means for a class_graph Object}
\usage{
\method{class_means}{class_graph}(x, X)
}
\arguments{
\item{x}{A class_graph object.}

\item{X}{A numeric matrix or data frame where rows represent observations and columns represent features.}
}
\value{
A matrix where each row contains the mean values for a class, with row names corresponding to class labels.
}
\description{
Compute Class Means for a class_graph Object
}
\examples{
# Using iris dataset
data(iris)
X <- as.matrix(iris[, 1:4])
cg <- class_graph(iris$Species)
means <- class_means(cg, X)

}
