% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{class_means}
\alias{class_means}
\title{Class Means}
\usage{
class_means(x, ...)
}
\arguments{
\item{x}{An object (e.g., class_graph).}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A matrix or data frame representing the means of each class, the structure of which depends on the input object's class.
}
\description{
A generic function to compute the mean of each class.
}
