% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commute_time.R
\name{commute_time_distance}
\alias{commute_time_distance}
\title{Compute the commute-time distance between nodes in a graph}
\usage{
commute_time_distance(A, ncomp = nrow(A) - 1)
}
\arguments{
\item{A}{A symmetric, non-negative matrix representing the adjacency matrix of the graph}

\item{ncomp}{Integer, number of components to use in the computation, default is (nrow(A) - 1)}
}
\value{
A list with the following components:
  \item{eigenvectors}{Matrix, eigenvectors of the matrix M}
  \item{eigenvalues}{Vector, eigenvalues of the matrix M}
  \item{cds}{Matrix, the computed commute-time distances}
  \item{gap}{Numeric, the gap between the two largest eigenvalues}
  The returned object has class "commute_time" and "list".
}
\description{
This function computes the commute-time distance between nodes in a graph using either eigenvalue or
pseudoinverse methods.
}
\examples{
# Create an example adjacency matrix
A <- matrix(c(0, 1, 1, 0,
             1, 0, 1, 1,
             1, 1, 0, 1,
             0, 1, 1, 0), nrow = 4, byrow = TRUE)

# Compute the commute-time distance
result <- commute_time_distance(A)

}
