% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{cross_spatial_adjacency}
\alias{cross_spatial_adjacency}
\title{cross_spatial_adjacency}
\usage{
cross_spatial_adjacency(coord_mat1, coord_mat2, dthresh = 1.42, nnk = 27,
  weight_mode = c("binary", "heat"), sigma = dthresh/2, normalized = TRUE)
}
\arguments{
\item{coord_mat1}{the first coordinate matrix}

\item{coord_mat2}{the second coordinate matrix}

\item{dthresh}{the distance threshold}

\item{nnk}{the maximum number of neighbors to include}

\item{sigma}{the bandwidth of the heat kernel if weight_mode == "heat"}

\item{normalized}{make row elements sum to 1}
}
\description{
cross_spatial_adjacency
}

