% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{cross_spatial_adjacency}
\alias{cross_spatial_adjacency}
\title{Cross Spatial Adjacency}
\usage{
cross_spatial_adjacency(
  coord_mat1,
  coord_mat2,
  dthresh = sigma * 3,
  nnk = 27,
  weight_mode = c("binary", "heat"),
  sigma = 5,
  normalized = TRUE
)
}
\arguments{
\item{coord_mat1}{A matrix with the spatial coordinates of the first set of data points, where each row represents a point and each column represents a coordinate dimension.}

\item{coord_mat2}{A matrix with the spatial coordinates of the second set of data points, where each row represents a point and each column represents a coordinate dimension.}

\item{dthresh}{The distance threshold for nearest neighbors (default: sigma * 3).}

\item{nnk}{The number of nearest neighbors to consider (default: 27).}

\item{weight_mode}{The mode to use for weighting the adjacency matrix, either "binary" or "heat" (default: "binary").}

\item{sigma}{The bandwidth for heat kernel weights (default: 5).}

\item{normalized}{A logical value indicating whether to normalize the adjacency matrix (default: TRUE).}
}
\value{
A sparse cross adjacency matrix with weighted spatial relationships between the two sets of points.
}
\description{
Constructs a cross spatial adjacency matrix between two sets of points, where weights are determined by spatial relationships.
}
\examples{
coord_mat1 <- as.matrix(expand.grid(x=1:5, y=1:5, z=1:5))
coord_mat2 <- as.matrix(expand.grid(x=6:10, y=6:10, z=6:10))
csa <- cross_spatial_adjacency(coord_mat1, coord_mat2, nnk=3, weight_mode="binary", sigma=5, normalized=TRUE)

}
