% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{cross_weighted_spatial_adjacency}
\alias{cross_weighted_spatial_adjacency}
\title{cross_weighted_spatial_adjacency}
\usage{
cross_weighted_spatial_adjacency(coord_mat1, coord_mat2, feature_mat1,
  feature_mat2, wsigma = 0.73, alpha = 0.5, nnk = 27,
  weight_mode = c("binary", "heat"), sigma = 1, dthresh = sigma * 2.5,
  normalized = TRUE)
}
\arguments{
\item{coord_mat1}{the first coordinate matrix (the query)}

\item{coord_mat2}{the second coordinate matrix (the reference)}

\item{feature_mat1}{the first feature matrix}

\item{feature_mat2}{the second feature matrix}

\item{wsigma}{the sigma for the feature heat kernel}

\item{alpha}{the mixing weight for the spatial distance (1=all spatial, 0=all feature)}

\item{dthresh}{the threshold for the spatial distance}

\item{normalized}{whether to normalize the rows to sum to 1}
}
\description{
cross_weighted_spatial_adjacency
}
