% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{discriminating_distance}
\alias{discriminating_distance}
\title{Compute Discriminating Distance for Similarity Graph}
\usage{
discriminating_distance(X, k = length(labels)/2, sigma, labels)
}
\arguments{
\item{X}{A numeric matrix or data frame containing the data points.}

\item{k}{An integer representing the number of nearest neighbors to consider.}

\item{sigma}{A numeric value representing the scaling factor for the heat kernel. If not provided, it will be estimated.}

\item{labels}{A factor or numeric vector containing the class labels for each data point.}
}
\value{
A discriminating distance matrix in the form of a numeric matrix.
}
\description{
This function computes a discriminating distance matrix for the similarity graph based on the class labels.
It adjusts the similarity graph by modifying the weights within and between classes, making it more suitable for
tasks like classification and clustering.
}
\examples{
\dontrun{
X <- matrix(rnorm(100*100), 100, 100)
labels <- factor(rep(1:5, each=20))
sigma <- 0.7
D <- discriminating_distance(X, k=length(labels)/2, sigma, labels)
}

}
