% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{discriminating_simililarity}
\alias{discriminating_simililarity}
\title{Compute similarity graph weighted by class structure}
\usage{
discriminating_simililarity(
  X,
  k = length(labels)/2,
  sigma,
  cg,
  threshold = 0.01
)
}
\arguments{
\item{cg}{}
}
\description{
Compute similarity graph weighted by class structure
}
\examples{

X <- matrix(rnorm(100*100), 100,100)
labels <- factor(rep(1:5, each=20))
cg <- class_graph(labels)
sigma <- .7

}
