% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{dist_to_sim}
\alias{dist_to_sim}
\title{Convert distance to similarity}
\usage{
dist_to_sim(x, ...)
}
\arguments{
\item{x}{An object representing a distance matrix.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A similarity matrix.
}
\description{
Convert distance to similarity
}
