% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{edges.neighbor_graph}
\alias{edges.neighbor_graph}
\title{Extract edges from neighbor_graph object}
\usage{
\method{edges}{neighbor_graph}(x)
}
\arguments{
\item{x}{A neighbor_graph object.}
}
\value{
A two-column matrix containing the edges, where each row represents an edge between two nodes.
}
\description{
Retrieve the edges of a neighbor_graph object.
}
\examples{
adj_matrix <- Matrix::Matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 0), 
                            nrow = 3, byrow = TRUE, sparse = TRUE)
ng <- neighbor_graph(adj_matrix)
edge_list <- edges(ng)

}
