% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{estimate_sigma}
\alias{estimate_sigma}
\title{estimate a reasonable bandwith parameter, sigma, for the heat kernel}
\usage{
estimate_sigma(X, prop = 0.25, nsamples = 500, normalized = FALSE)
}
\arguments{
\item{X}{the data matrix, samples are rows, features are columns}

\item{prop}{the quantile of the frequency distribution of distances used to determine bandwidth parameter.}

\item{nsamples}{the number of samples to draw from the data matrix}
}
\description{
bandwidth estimation
}
