% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{factor_sim}
\alias{factor_sim}
\title{factor_sim}
\usage{
factor_sim(des, method = c("Jaccard", "Rogers", "simple matching", "Dice"))
}
\description{
compute similarity matrix from a set of \code{factor}s in a \code{data.frame}
}
