% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{graph_weights}
\alias{graph_weights}
\title{Convert a data matrix with n instances and p features to an n-by-n adjacency matrix}
\usage{
graph_weights(X, k = 5, neighbor_mode = c("knn", "supervised",
  "knearest_misses", "epsilon"), weight_mode = c("heat", "normalized",
  "binary", "euclidean", "cosine"), type = c("normal", "mutual", "asym"),
  sigma, eps = NULL, labels = NULL)
}
\arguments{
\item{X}{the data matrix, where each row is an instance and each column is a variable. Similarity is computed over instances.}

\item{k}{number of neighbors}

\item{neighbor_mode}{the method for assigning weights to neighbors, either "supervised", "knn", "knearest_misses", or "epsilon"}

\item{weight_mode}{binary (1 if neighbor, 0 otherwise),'heat', 'normalized', 'euclidean', or 'cosine'}

\item{type}{the nearest neighbor policy, one of: normal, mutual, asym.}

\item{sigma}{parameter for heat kernel \code{exp(-dist/(2*sigma^2))}}

\item{eps}{the neighborhood radius when neighbor_mode is `epsilon`}

\item{labels}{the class of the categories when \code{weight_mode} is \code{supervised}, supplied as a \code{factor} with \code{nrow(labels) == nrow(X)}}
}
\description{
Convert a data matrix with n instances and p features to an n-by-n adjacency matrix
}
\examples{

X <- matrix(rnorm(20*10), 20, 10)
sm <- graph_weights(X, neighbor_mode="knn",k=3)

labels <- factor(rep(letters[1:4],5))
sm2 <- graph_weights(X, neighbor_mode="supervised",k=3, labels=labels)

sm3 <- graph_weights(X, neighbor_mode="knearest_misses",k=3, labels=labels, weight_mode="binary")
}
