% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sim.R
\name{label_matrix}
\alias{label_matrix}
\title{label_matrix}
\usage{
label_matrix(
  a,
  b,
  type = c("s", "d"),
  return_matrix = TRUE,
  simfun = NULL,
  dim1 = length(a),
  dim2 = length(b)
)
}
\arguments{
\item{a}{the first categorical label vector}

\item{b}{the second categorical label vector}

\item{type}{construct a a similarity ('s') or distance ('d') matrix.}

\item{return_matrix}{return as a sparse matrix or a triplet.}

\item{the}{length of the second dimension}
}
\description{
label_matrix
}
