% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{laplacian.neighbor_graph}
\alias{laplacian.neighbor_graph}
\title{Laplacian Matrix for neighbor_graph Objects}
\usage{
\method{laplacian}{neighbor_graph}(x, normalized = FALSE)
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{normalized}{A logical value indicating whether the normalized Laplacian should be computed.}
}
\value{
A Matrix object representing the Laplacian matrix.
}
\description{
Compute the Laplacian matrix of a neighbor_graph object.
}
