% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{laplacian}
\alias{laplacian}
\title{compute graph laplacian of a weight matrix}
\usage{
laplacian(x, ...)
}
\arguments{
\item{x}{the weight matrix}

\item{...}{extra args}
}
\description{
compute graph laplacian of a weight matrix
}
