% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{nclasses.class_graph}
\alias{nclasses.class_graph}
\title{Number of Classes for class_graph Objects}
\usage{
\method{nclasses}{class_graph}(x)
}
\arguments{
\item{x}{A class_graph object.}
}
\value{
The number of classes in the class_graph.
}
\description{
Compute the number of classes in a class_graph object.
}
