% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{neighbor_graph.Matrix}
\alias{neighbor_graph.Matrix}
\title{Neighbor Graph for Matrix Objects}
\usage{
\method{neighbor_graph}{Matrix}(x, params = list(), type = NULL, classes = NULL, ...)
}
\arguments{
\item{x}{A Matrix object.}

\item{params}{A list of parameters.}

\item{type}{A character string specifying the type of graph.}

\item{classes}{A character vector specifying additional classes for the object.}

\item{...}{Additional arguments.}
}
\value{
A neighbor_graph object.
}
\description{
Create a neighbor_graph object from a Matrix object.
}
