% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{neighbor_graph}
\alias{neighbor_graph}
\title{Neighbor Graph}
\usage{
neighbor_graph(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A neighbor_graph object, the structure of which depends on the input object's class.
}
\description{
A generic function to create a neighbor_graph object.
}
