% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{neighbors.neighbor_graph}
\alias{neighbors.neighbor_graph}
\title{Neighbors for neighbor_graph Objects}
\usage{
\method{neighbors}{neighbor_graph}(x, i)
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{i}{An integer specifying the index of the node for which neighbors should be retrieved.}
}
\value{
A list of neighbors.
}
\description{
Retrieve the neighbors of a specific node or all nodes in a neighbor_graph object.
}
