% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_graph.R
\name{node_density.neighbor_graph}
\alias{node_density.neighbor_graph}
\title{Node Density for neighbor_graph Objects}
\usage{
\method{node_density}{neighbor_graph}(x, X)
}
\arguments{
\item{x}{A neighbor_graph object.}

\item{X}{A data matrix containing the data points.}
}
\value{
A numeric vector containing the node densities.
}
\description{
Compute the node density for a neighbor_graph object.
}
