% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{non_neighbors}
\alias{non_neighbors}
\title{Get Indices of Non-neighbors of a Node}
\usage{
non_neighbors(x, ...)
}
\arguments{
\item{x}{A neighbor graph object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A numeric vector of node indices that are not neighbors of the given node.
}
\description{
Retrieve the indices of nodes that are not neighbors of a specified node.
}
