% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{normalize_adjacency}
\alias{normalize_adjacency}
\title{normalize an adjacency matrix}
\usage{
normalize_adjacency(sm, symmetric = TRUE)
}
\arguments{
\item{sm}{the adjacency matrix}

\item{symmetric}{whether to symmetrize after normalizing}
}
\description{
normalize an adjacency matrix
}
