% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generic.R
\name{nvertices}
\alias{nvertices}
\title{Number of Vertices in Graph-like Objects}
\usage{
nvertices(x, ...)
}
\arguments{
\item{x}{An object with a neighborhood structure.}

\item{...}{Additional arguments passed to methods (currently ignored).}
}
\value{
An integer representing the number of vertices in the graph.
}
\description{
Retrieve the number of vertices in a graph-like object. This function is a generic method
that works with various types of graph representations.
}
\examples{
# Create a small example graph
adj_matrix <- matrix(c(0, 1, 1, 0, 1, 0, 1, 0, 0), nrow = 3, byrow = TRUE)
ng <- neighbor_graph(adj_matrix)
nvertices(ng) # Should return 3

}
