% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{repulsion_graph}
\alias{repulsion_graph}
\title{construct a "repulsion graph"}
\usage{
repulsion_graph(
  labels,
  W,
  X = NULL,
  sigma = 10,
  method = c("binary", "weighted")
)
}
\arguments{
\item{labels}{a set of labels defining the classes}

\item{W}{the neighborhood graph (weighted or binary)}

\item{X}{the original data matrix where \code{nrow(X) == nrow(W)}}

\item{sigma}{bandwidth of the similarity function}
}
\description{
construct a "repulsion graph"
}
