% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repulsion.R
\name{repulsion_graph}
\alias{repulsion_graph}
\title{construct a "repulsion graph"}
\usage{
repulsion_graph(
  W,
  cg,
  method = c("binary", "weighted"),
  threshold = 0,
  norm_fac = 1
)
}
\arguments{
\item{W}{the adjacency matrix (weighted or binary)}

\item{cg}{the class graph indicating the pairwise relationships among node classes}
}
\description{
construct a "repulsion graph"
}
\examples{

X <- matrix(rnorm(100*100), 100,100)
labels <- factor(rep(1:5, each=20))
cg <- class_graph(labels)
W <- graph_weights(X, k=10)
R <- repulsion_graph(W, cg, method="weighted")

data(iris)
X2 <- as.matrix(iris[,1:4])
labels <- iris[,5]
cg <- class_graph(labels)
W <- graph_weights(X2, k=6)
R <- repulsion_graph(W, cg, method="weighted")
}
