% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{spatial_adjacency}
\alias{spatial_adjacency}
\title{Compute the Spatial Adjacency Matrix for a Coordinate Matrix}
\usage{
spatial_adjacency(
  coord_mat,
  dthresh = sigma * 3,
  nnk = 27,
  weight_mode = c("binary", "heat"),
  sigma = 5,
  include_diagonal = TRUE,
  normalized = TRUE,
  stochastic = FALSE
)
}
\arguments{
\item{coord_mat}{A numeric matrix representing the spatial coordinates}

\item{dthresh}{Numeric, the distance threshold defining the radius of the neighborhood (default is sigma*3)}

\item{nnk}{Integer, the maximum number of neighbors to include in each spatial neighborhood (default is 27)}

\item{weight_mode}{Character, the mode for computing weights, either "binary" or "heat" (default is "binary")}

\item{sigma}{Numeric, the bandwidth of the heat kernel if weight_mode == "heat" (default is 5)}

\item{include_diagonal}{Logical, whether to assign 1 to diagonal elements (default is TRUE)}

\item{normalized}{Logical, whether to make row elements sum to 1 (default is TRUE)}

\item{stochastic}{Logical, whether to make column elements also sum to 1 (only relevant if normalized == TRUE) (default is FALSE)}
}
\value{
A sparse symmetric matrix representing the computed spatial adjacency
}
\description{
Computes the spatial adjacency matrix for a given coordinate matrix using specified parameters.
}
\details{
The function first computes the spatial adjacency matrix using the specified parameters,
and then normalizes the matrix to make it row-stochastic.
}
\examples{
coord_mat = as.matrix(expand.grid(x=1:6, y=1:6))
sa <- spatial_adjacency(coord_mat)
}
