% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{spatial_adjacency}
\alias{spatial_adjacency}
\title{spatial_adjacency}
\usage{
spatial_adjacency(coord_mat, dthresh = 1.42, nnk = 27,
  weight_mode = c("binary", "heat"), sigma = dthresh/2,
  include_diagonal = TRUE, normalized = TRUE, stochastic = FALSE)
}
\arguments{
\item{coord_mat}{the matrix of spatial coordinates}

\item{dthresh}{the distance threshold defining the radius of the neighborhood}

\item{nnk}{the maximum number of neighbors to include in each spatial neighborhood}

\item{weight_mode}{a binary or heat kernel}

\item{sigma}{the bandwidth of the heat kernel if weight_mode == "heat"}

\item{include_diagonal}{diagonal elements assigned 1}

\item{normalized}{make row elements sum to 1}

\item{stochastic}{make column elements also sum to 1 (only relevant if \code{normalized == TRUE})}
}
\description{
spatial_adjacency
}
\examples{

coord_mat = as.matrix(expand.grid(x=1:6, y=1:6))
sa <- spatial_adjacency(coord_mat)
}
