% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_constraints.R
\name{spatial_constraints}
\alias{spatial_constraints}
\title{spatial constraints}
\usage{
spatial_constraints(
  coords,
  nblocks = 1,
  sigma_within = 5,
  sigma_between = 1,
  shrinkage_factor = 0.1,
  nnk_within = 27,
  nnk_between = 1,
  weight_mode_within = "heat",
  weight_mode_between = "binary",
  variable_weights = 1,
  verbose = FALSE
)
}
\arguments{
\item{coords}{the spatial coordinates as a matrix with rows as objects and columns as dimensions; or as a
list of matrices where each element of the list contains the coordinate for a block.}

\item{nblocks}{the number of coordinate blocks}

\item{sigma_within}{the bandwidth of the within-block smoother}

\item{sigma_between}{the bandwidth of the between-block smoother}

\item{shrinkage_factor}{the amount of shrinkage towards the spatial block average}

\item{nnk_within}{the maximum number of nearest neighbors for within block smoother}

\item{nnk_between}{the maximum number of nearest neighbors for between block smoother}

\item{weight_mode_within}{the within block nearest neighbor weight_mode ("heat" or "binary")}

\item{weight_mode_between}{the between block nearest neighbor weight_mode ("heat" or "binary")}

\item{variable_weights}{a set of per variable weights}
}
\description{
construct a sparse matrix of spatial constraints for a set of data blocks
}
