% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_weights.R
\name{temporal_autocor}
\alias{temporal_autocor}
\title{temporal autocorrelation}
\usage{
temporal_autocor(X, window = 3, inverse = FALSE)
}
\arguments{
\item{X}{the data matrix, columns are variables, rows are time-points}

\item{window}{th number of time-points included in the auto-correlation band}

\item{inverse}{whether to return the inverse correlation matrix uisng `corpcor::invcor.shrink`}
}
\description{
Compute a sparse temporal autocorrelation matrix for a temporal window
}
