% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_weights.R
\name{weighted_knn}
\alias{weighted_knn}
\title{weighted_knn}
\usage{
weighted_knn(
  X,
  k = 5,
  FUN = heat_kernel,
  type = c("normal", "mutual", "asym"),
  return_triplet = FALSE,
  ...
)
}
\arguments{
\item{X}{the data matrix, where rows are instances and columns are features}

\item{k}{the number of nearest neighbors}

\item{FUN}{the function used to convert euclidean distances to similarities}

\item{type}{whether to compute 'normal' or 'mutual' k-nearest neighbors}
}
\description{
weighted_knn
}
\examples{

X <- matrix(rnorm(10*10), 10,10)
w <- weighted_knn(X,k=5)

}
