% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_weights.R
\name{weighted_spatial_adjacency}
\alias{weighted_spatial_adjacency}
\title{construct a spatial adjacency matrix weighted by a secondary feature matrix}
\usage{
weighted_spatial_adjacency(
  coord_mat,
  feature_mat,
  wsigma = 0.73,
  alpha = 0.5,
  nnk = 27,
  weight_mode = c("binary", "heat"),
  sigma = 1,
  dthresh = sigma * 2.5,
  include_diagonal = TRUE,
  normalized = FALSE,
  stochastic = FALSE
)
}
\arguments{
\item{coord_mat}{the coordinate matrix}

\item{feature_mat}{the feature matrix, where: (\code{nrow(feature_mat) == nrow(coord_mat)})}

\item{wsigma}{the sigma for the feature heat kernel}

\item{alpha}{the mixing weight for the spatial distance (1=all spatial weighting, 0=all feature weighting)}

\item{nnk}{the maximum number of spatial nearest neighbors to include}

\item{dthresh}{the threshold for the spatial distance}

\item{normalized}{whether to normalize the rows to sum to 1}
}
\description{
construct a spatial adjacency matrix weighted by a secondary feature matrix
}
\examples{

coord_mat <- as.matrix(expand.grid(x=1:9, y=1:9, z=1:9))
fmat <- matrix(rnorm(nrow(coord_mat)*100), nrow(coord_mat), 100)
wsa1 <- weighted_spatial_adjacency(coord_mat, fmat, nnk=3, weight_mode="binary", alpha=1, stochastic=TRUE)
wsa2 <- weighted_spatial_adjacency(coord_mat, fmat, nnk=27, weight_mode="heat", alpha=0, stochastic=TRUE, sigma=2.5)
}
