% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_graph.R
\name{within_class_neighbors.class_graph}
\alias{within_class_neighbors.class_graph}
\title{Within-Class Neighbors for class_graph Objects}
\usage{
\method{within_class_neighbors}{class_graph}(x, ng)
}
\arguments{
\item{x}{A class_graph object.}

\item{ng}{A neighbor graph object.}
}
\value{
A neighbor_graph object representing the within-class neighbors of the input class_graph.
}
\description{
Compute the within-class neighbors of a class_graph object.
}
